Create separate mutation for password reset
